# docs/swagger.yaml
definitions:
  structs.Player:
    properties:
      birth:
        type: string
      city:
        type: string
      country:
        type: string
      height:
        type: string
      id:
        type: integer
      idteam:
        type: integer
      name:
        type: string
    type: object
  structs.Team:
    properties:
      city:
        type: string
      country:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: Essa API permite que os usuários gerenciem times e jogadores de futebol.
    Os usuários podem inserir novos times e jogadores, atualizar informações existentes
    e consultar detalhes.
  title: Squads API
paths:
  /api/v1/player/delete/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a player by ID
      parameters:
      - description: Player ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: Delete a player by ID
      tags:
      - Players
...
  /api/v1/player/getbycountry/{country}:
    get:
      consumes:
      - application/json
      description: Get players by country
      parameters:
      - description: Country Name
        in: path
        name: country
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/structs.Player'
            type: array
      summary: Get players by country
      tags:
      - Players
  /api/v1/player/getbyid/{id}:
    get:
      consumes:
      - application/json
      description: Get a player by ID
      parameters:
      - description: Player ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/structs.Player'
      summary: Get a player by ID
      tags:
      - Players
  /api/v1/player/getbyidteam/{idteam}:
    get:
      consumes:
      - application/json
      description: Get players by team ID
      parameters:
      - description: IdTeam
        in: path
        name: idteam
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/structs.Player'
            type: array
      summary: Get players by team ID
      tags:
      - Players
  /api/v1/player/getbyname/{name}:
    get:
      consumes:
      - application/json
      description: Get players by name
      parameters:
      - description: Player Name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/structs.Player'
            type: array
      summary: Get players by name
      tags:
      - Players
  /api/v1/player/insert:
    post:
      consumes:
      - application/json
      description: Insert a player
      parameters:
      - description: Player Name
        in: query
        name: name
        required: true
        type: string
      - description: Id Team
        in: query
        name: idteam
        required: true
        type: string
      - description: City
        in: query
        name: city
        required: true
        type: string
      - description: Country
        in: query
        name: country
        required: true
        type: string
      - description: Birth
        in: query
        name: birth
        required: true
        type: string
      - description: Height
        in: query
        name: height
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/structs.Player'
      summary: Insert a player
      tags:
      - Players
  /api/v1/player/update:
    put:
      consumes:
      - application/json
      description: Update a player
      parameters:
      - description: ID Player
        in: query
        name: id
        required: true
        type: integer
      - description: Player Name
        in: query
        name: name
        required: true
        type: string
      - description: Id Team
        in: query
        name: idteam
        required: true
        type: integer
      - description: City
        in: query
        name: city
        required: true
        type: string
      - description: Country
        in: query
        name: country
        required: true
        type: string
      - description: Birth
        in: query
        name: birth
        required: true
        type: string
      - description: Height
        in: query
        name: height
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/structs.Player'
      summary: Update a player
      tags:
      - Players
...
  /api/v1/team/getbycountry/{country}:
    get:
      consumes:
      - application/json
      description: Get teams by country
      parameters:
      - description: Country Name
        in: path
        name: country
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/structs.Team'
            type: array
      summary: Get teams by country
      tags:
      - Teams
  /api/v1/team/getbyid/{id}:
    get:
      consumes:
      - application/json
      description: Get a team by ID
      parameters:
      - description: Team ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/structs.Team'
      summary: Get a team by ID
      tags:
      - Teams
  /api/v1/team/getbyname/{name}:
    get:
      consumes:
      - application/json
      description: Get teams by name
      parameters:
      - description: Team Name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/structs.Team'
            type: array
      summary: Get teams by name
      tags:
      - Teams
  /api/v1/team/insert:
    post:
      consumes:
      - application/json
      description: Insert a team
      parameters:
      - description: Team Name
        in: query
        name: name
        required: true
        type: string
      - description: City
        in: query
        name: city
        required: true
        type: string
      - description: Country
        in: query
        name: country
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/structs.Team'
      summary: Insert a team
      tags:
      - Teams
  /api/v1/team/update:
    put:
      consumes:
      - application/json
      description: Update a team
      parameters:
      - description: ID Team
        in: query
        name: id
        required: true
        type: integer
      - description: Team Name
        in: query
        name: name
        required: true
        type: string
      - description: City
        in: query
        name: city
        required: true
        type: string
      - description: Country
        in: query
        name: country
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/structs.Team'
      summary: Update a team
      tags:
      - Teams
schemes:
- http
swagger: "2.0"